+++ R/bagged.R
@@ -51,12 +51,6 @@ bagged <- function(resp, trt, subgr, covars = NULL, data,
   res <- do.call(fitMods, args = all.args)
   crit <- qnorm(1 - level / 2)
   nSub <- length(subgr)
-  ## extract overall estimates
-  tmpO <- c(res$fitOverall$ests$ests,
-            sqrt(res$fitOverall$ests$ests_vc))
-  overallEff <- c(tmpO[1],
-                  tmpO[1] - crit * tmpO[2],
-                  tmpO[1] + crit * tmpO[2])
   
   # Obtain Treatment*subgroup interactions
   results <- do.call(rbind, lapply(res$subgroups, function(ii){
@@ -100,7 +94,7 @@ bagged <- function(resp, trt, subgr, covars = NULL, data,
   }
   ## Analyze bootstrap samples -------------------------------------------------
   if (mc.cores > 1){
-    boot_results <- parallel::mclapply(numberhits, function(x){
+    boot_res <- parallel::mclapply(numberhits, function(x){
                                        fitMods_boot(x,
                                                     ordered_data = ordered_data, 
                                                     resp = resp, trt = trt,
@@ -111,7 +105,7 @@ bagged <- function(resp, trt, subgr, covars = NULL, data,
                                                     dots. = dots.)
     }, mc.cores = mc.cores)
   } else {
-    boot_results <- lapply(numberhits, function(x){
+    boot_res <- lapply(numberhits, function(x){
       fitMods_boot(x,
                    ordered_data = ordered_data, 
                    resp = resp, trt = trt,
@@ -124,85 +118,55 @@ bagged <- function(resp, trt, subgr, covars = NULL, data,
   }
   
   ## Get the estimates of the interaction with the minimum SBC
-  boot_results_ext <- t(sapply(boot_results, 
-                               function(x) x[which.min(x[, select.by]), 
-                                             c("subgroup", "estimate", "SubgrEff", "ComplEff")]))
-  ## Transform it into a dataframe for better handling
-  boot_results_ext <- data.frame(subgroup = unlist(boot_results_ext[, 1]), 
-                                 estimate = unlist(boot_results_ext[, 2]), 
-                                 SubgrEff = unlist(boot_results_ext[, 3]), 
-                                 ComplEff = unlist(boot_results_ext[, 4]))
+  boot_results_ext <- do.call(rbind,lapply(boot_res, 
+                               function(x) {
+                                  x[which.min(x[, select.by]), 
+                                             c("subgroup", "estimate", "SubgrEff", "ComplEff")]
+                               }))
+  rownames(boot_results_ext) = NULL
+  boot_invalid = sum(sapply(boot_res, is.null))/B*100
+  numberhits = numberhits[!sapply(boot_res, is.null)] # Keep only valid bootstrap samples
+  boot_res = boot_res[!sapply(boot_res, is.null)]     # Keep only valid bootstrap samples
+  InteractionEst = baggingSoupr("estimate", results, boot_res, boot_results_ext, subgr, numberhits, select.by, N)
+  SubgrEff       = baggingSoupr("SubgrEff", results, boot_res, boot_results_ext, subgr, numberhits, select.by, N)
+  ComplEff       = baggingSoupr("ComplEff", results, boot_res, boot_results_ext, subgr, numberhits, select.by, N)
+  postpr = unlist(InteractionEst["percent_selected"])
+  names(postpr) <- c(subgr, "overall")
   
-  InteractionEst = baggingSoupr("estimate", results, boot_results, boot_results_ext, subgr, numberhits, select.by, B, N, level)
-  SubgrEff       = baggingSoupr("SubgrEff", results, boot_results, boot_results_ext, subgr, numberhits, select.by, B, N, level)
-  ComplEff       = baggingSoupr("ComplEff", results, boot_results, boot_results_ext, subgr, numberhits, select.by, B, N, level)
+  boot_results <- list(selected = names(which(selected == 1)),
+                       unadjusted = results,
+                       percSel = postpr[-(nSub+1)],
+                       B = B, boot_invalid = boot_invalid,
+                       InteractionEst = InteractionEst, 
+                       SubgrEff = SubgrEff, 
+                       ComplEff = ComplEff)
   
   not_selected = paste(subgr[(which(InteractionEst$percent_selected[-(nSub+1)] == 0))])
   if (length(not_selected) > 0) message(paste0("The subgroup(s) ", paste(not_selected, collapse = ", "),
                                                " were not selected in the bootstrap samples and a corrected estimate is not available. Returning NA."))
   if(any(is.nan(InteractionEst$se.bagg_red)) & B < 2000) message("The variance for the bootstrap estimate of one or more subgroups could not be calculated because they were not selected in sufficient bootstrap samples. This may be due to a small number of bootstrap samples (B) or simply that the subgroup is not predictive.")
   
-  postpr = unlist(InteractionEst["percent_selected"])
   model_fits <- results 
   rownames(model_fits) <- NULL
+  
   ## produce estimates for subgroups
-  lst <- vector("list", nSub)
-  for(i in 1:nSub){
-    est <- numeric(3)
-    resFit <- res$fitmods[[i]]$ests
-    tmp1 <- c(SubgrEff[i, "bagg_red"], SubgrEff[i, "se.bagg_red"])
-    tmp2 <- c(ComplEff[i, "bagg_red"], ComplEff[i, "se.bagg_red"])
-    tmp3 <- c(InteractionEst[i, "bagg_red"], InteractionEst[i, "se.bagg_red"])
-    est <- c(tmp1[1], tmp2[1], tmp3[1])
-    ub <- est+crit*c(tmp1[2], tmp2[2], tmp3[2])
-    lb <- est-crit*c(tmp1[2], tmp2[2], tmp3[2])
-    lst[[i]] <- list(resFit$pval[2],
-                     est[1:2], ub[1:2], lb[1:2],
-                     est[3], ub[3], lb[3])
-  }
-  
-  pvals <- sapply(lst, function(x) x[[1]])
-  trt_eff <- c(sapply(lst, function(x) x[[2]]))
-  trt_ub <- c(sapply(lst, function(x) x[[3]]))
-  trt_lb <- c(sapply(lst, function(x) x[[4]]))
-  ia <- sapply(lst, function(x) x[[5]])
-  ia_ub <- sapply(lst, function(x) x[[6]])
-  ia_lb <- sapply(lst, function(x) x[[7]])
-  fitmods <- lapply(res$fitmods, function(x) x$model)
-  fitmods[["overall"]] <- res$fitOverall$model
-  
-  
-  Group <- sapply(res$fitmods, function(x) x$subgrNam)
-  cmpl <- rep(c("Subgroup", "Complement"), nSub)
+  cis = getConfints(res, data, type = "bagged", nSub, level, postpr, boot_results)
   
   obj <- list()
-  obj$fitmods <- fitmods
-  obj$trtEff <- data.frame(Group = rep(Group, each=2),
-                           Subset = cmpl,
-                           LB = trt_lb, trtEff = trt_eff,
-                           UB = trt_ub)
-  obj$trtEffDiff <- data.frame(Group = Group,
-                               LB = ia_lb, trtEffDiff = ia,
-                               UB = ia_ub)
-  rownames(obj$trtEffDiff) <- NULL
+  obj$fitMods <- res
+  obj$trtEff <- cis$trtEff
+  obj$trtEffDiff <- cis$trtEffDiff
   obj$n <- nrow(data)
   obj$subgroups <- res$subgroups
   obj$GroupSizes <- sizes
   obj$level <- level
-  obj$overall <- overallEff 
-  names(postpr) <- c(subgr, "overall")
+  obj$overall <- cis$overall 
   obj$post.weights <- postpr
-  names(pvals) <- subgr
-  obj$pvals <- pvals
+  obj$pvals <- cis$pvals
   obj$type <- "bagged"
   obj$fitfunc <- fitfunc
-  obj$boot_results <- list(selected = names(which(selected == 1)),
-                           unadjusted = results,
-                           percSel = postpr[-(nSub+1)],
-                           B = B,
-                           InteractionEst = InteractionEst, 
-                           SubgrEff = SubgrEff, 
-                           ComplEff = ComplEff)
+  obj$data <- data
+  obj$boot_results <- boot_results
   class(obj) <- "subtee"
   obj
 }
@@ -210,7 +174,8 @@ bagged <- function(resp, trt, subgr, covars = NULL, data,
 
 
 baggingSoupr = function(target, results, boot_results, boot_results_ext, 
-                        subgr, numberhits, select.by, B, N, level){
+                        subgr, numberhits, select.by, N){
+  B = length(boot_results)
   ## Calculate mean_boot_estimate pstar mean_u_boot_estimate
   boot_estimate <- sapply(boot_results, function(x) x[[target]])
   mean_boot_estimate <- rowMeans(boot_estimate)
@@ -314,9 +279,6 @@ baggingSoupr = function(target, results, boot_results, boot_results_ext,
 }
 
 
-
-
-
 fitMods_boot = function(numberhits, ordered_data, 
                         resp, trt,
                         subgr, covars, 
@@ -326,33 +288,63 @@ fitMods_boot = function(numberhits, ordered_data,
                         dots.) {
   data_boot <- ordered_data[which(numberhits > 0), ]
   ww <- numberhits[which(numberhits > 0)]
-  args <- list(resp = resp, trt = trt,
-               subgr = subgr, covars = covars, 
-               data = data_boot, fitfunc = fitfunc, 
-               event = event, exposure = exposure,
-               weights = ww)
-  all.args <- c(args, dots.)
-  res <- do.call(fitMods, args = all.args)
-  results <- do.call(rbind, lapply(res$subgroups, function(ii){
-    x <- res$fitmods[[ii]]$model
-    p <- length(x$coef)
-    coef <- x$coefficients[p]
-    complEff <- res$fitmods[[ii]]$ests$ests[1]
-    se <- sqrt(diag(x$var)[p])
-    tmp <- c(coef, se, 1 - pchisq((coef/se)^2, 1), AIC(x), BIC(x), complEff, complEff + coef)
-    tmp
-  }))
-  rownames(results) <- res$subgroups
-  colnames(results) <- c("estimate", "std.error", "p.value", "AIC", "BIC", "ComplEff", "SubgrEff")
-  overall <- c(estimate = NA, 
-               std.error = NA,
-               p.value = NA,
-               AIC = res$fitOverall$ests$AIC,
-               BIC = res$fitOverall$ests$BIC,
-               ComplEff = res$fitOverall$ests$ests,
-               SubgrEff = res$fitOverall$ests$ests)
-  results <- rbind(results, overall)
-  selected <- 1*(results[, select.by] == min(results[, select.by]))
-  results <- data.frame(subgroup = c(res$subgroups, "none"), results, selected)
-  results
+  boot_valid = checkTrtSub_boot(data = data_boot, trt = trt, subgr = subgr, weights = ww)
+  if (boot_valid){
+    args <- list(resp = resp, trt = trt,
+                 subgr = subgr, covars = covars, 
+                 data = data_boot, fitfunc = fitfunc, 
+                 event = event, exposure = exposure,
+                 weights = ww)
+    all.args <- c(args, dots.)
+    res <- do.call(fitMods, args = all.args)
+    results <- do.call(rbind, lapply(res$subgroups, function(ii){
+      x <- res$fitmods[[ii]]$model
+      p <- length(x$coef)
+      coef <- x$coefficients[p]
+      complEff <- res$fitmods[[ii]]$ests$ests[1]
+      se <- sqrt(diag(x$var)[p])
+      tmp <- c(coef, se, 1 - pchisq((coef/se)^2, 1), AIC(x), BIC(x), complEff, complEff + coef)
+      tmp
+    }))
+    rownames(results) <- res$subgroups
+    colnames(results) <- c("estimate", "std.error", "p.value", "AIC", "BIC", "ComplEff", "SubgrEff")
+    overall <- c(estimate = NA, 
+                 std.error = NA,
+                 p.value = NA,
+                 AIC = res$fitOverall$ests$AIC,
+                 BIC = res$fitOverall$ests$BIC,
+                 ComplEff = res$fitOverall$ests$ests,
+                 SubgrEff = res$fitOverall$ests$ests)
+    results <- rbind(results, overall)
+    selected <- 1*(results[, select.by] == min(results[, select.by]))
+    results <- data.frame(subgroup = c(res$subgroups, "none"), results, selected, stringsAsFactors = FALSE)
+    results
+  }
 }
+
+checkTrtSub_boot <- function(data, trt, subgr, weights){
+  ## input checking
+  ind = TRUE
+  trttmp <- data[[trt]]
+  ## check for whether a subgroup defines the whole data-set and
+  ## whether for all subgroups and complements we have patients in control and trt
+  nSub <- length(subgr)
+  sizes <- complSizes <- NAs <- numeric(nSub)
+  for(i in 1:nSub){
+    tmpvar <- data[,subgr[i]]
+    len <- length(unique(tmpvar))
+    if(len == 1){
+      ind = FALSE
+    }
+    # Performs a table with weighted frequencies.
+    tochk = factor(paste0(tmpvar, trttmp), levels = c("00", "01", "10", "11"))
+    tab <- aggregate(weights, by = list(cat = tochk), sum, drop = FALSE)
+    tab$x <- ifelse(is.na(tab$x), 0, tab$x) #Nico's fix, because in new R version (>3.5.0), drop=FALSE creates NA instead of 0 for empty combinations
+    if(any(tab$x < 2 | is.na(tab$x) | (nrow(tab) < 4))){
+      ind = FALSE
+    }
+  }
+  ind
+}
+
+